adds loading spinner
